\name{aggregate_aa_per_sample}
\alias{aggregate_aa_per_sample}
\title{Aggregate amino acid counts per sample}
\usage{
aggregate_aa_per_sample(counts)
}
\arguments{
  \item{counts}{output of load_aa_and_meta}
}
\value{
Data frame with counts, wide-like
}
\description{
Aggregate amino acid counts per sample
}

