\docType{data}
\name{fill_scale}
\alias{fill_scale}
\title{Percent mutant fill scale}
\format{\preformatted{List of 14
 $ call      : language discrete_scale(aesthetics = aesthetic, scale_name = "manual", palette = pal,      name = "\% mutant", na.value = ..2, drop = FALSE)
 $ aesthetics: chr "fill"
 $ scale_name: chr "manual"
 $ palette   :function (n)  
 $ range     :Reference class 'DiscreteRange' [package "scales"] with 1 fields
  ..$ range: NULL
  ..and 15 methods, of which 3 are possibly relevant:
  ..  initialize, reset, train
 $ limits    : NULL
 $ na.value  : chr "#AAAAAA"
 $ expand    : list()
  ..- attr(*, "class")= chr "waiver"
 $ name      : chr "\% mutant"
 $ breaks    : list()
  ..- attr(*, "class")= chr "waiver"
 $ labels    : list()
  ..- attr(*, "class")= chr "waiver"
 $ legend    : NULL
 $ drop      : logi FALSE
 $ guide     : chr "legend"
 - attr(*, "class")= chr [1:3] "manual" "discrete" "scale"
}}
\usage{
fill_scale
}
\description{
Percent mutant fill scale
}
\keyword{datasets}

