\docType{data}
\name{genotype_specific_sites}
\alias{genotype_specific_sites}
\title{Pattern to match sites which use a genotype-specific correction}
\format{\preformatted{ chr "^(K65|D67)"
}}
\usage{
genotype_specific_sites
}
\description{
Pattern to match sites which use a genotype-specific
correction
}
\keyword{datasets}

