\name{get_direction}
\alias{get_direction}
\title{Determines whether a mutation should be read by the forward or reverse read}
\usage{
get_direction(s)
}
\arguments{
  \item{s}{Mutation string}
}
\value{
'forward' or 'reverse'
}
\description{
Determines whether a mutation should be read by the forward
or reverse read
}

