\name{load_aa_and_meta}
\alias{load_aa_and_meta}
\title{Load and merge amino acid counts and metadata}
\usage{
load_aa_and_meta(count_path = "output/prep_merged_classified.csv",
  meta_path = default_meta_path)
}
\arguments{
  \item{count_path}{path to counts}

  \item{meta_path}{path to metadata csv}
}
\value{
Data frame with merged results
}
\description{
Load and merge amino acid counts and metadata
}

