\name{load_counts_and_meta}
\alias{load_counts_and_meta}
\title{Load and merge mutation counts and metadata}
\usage{
load_counts_and_meta(count_path = "output/prep_merged_classified_counts.csv",
  meta_path = default_meta_path)
}
\arguments{
  \item{count_path}{Path to counts file}

  \item{meta_path}{Path to sample metadata}
}
\value{
Data frame with merged results
}
\description{
Load and merge mutation counts and metadata
}

