\docType{data}
\name{p_value_fill}
\alias{p_value_fill}
\title{Palette for filling cells colored by p-value}
\format{\preformatted{List of 14
 $ call      : language discrete_scale(aesthetics = aesthetic, scale_name = "manual", palette = pal,      name = "p-value\\n(FDR corrected)", na.value = ..3, drop = FALSE)
 $ aesthetics: chr "fill"
 $ scale_name: chr "manual"
 $ palette   :function (n)  
 $ range     :Reference class 'DiscreteRange' [package "scales"] with 1 fields
  ..$ range: NULL
  ..and 15 methods, of which 3 are possibly relevant:
  ..  initialize, reset, train
 $ limits    : NULL
 $ na.value  : chr "#AAAAAA"
 $ expand    : list()
  ..- attr(*, "class")= chr "waiver"
 $ name      : chr "p-value\\n(FDR corrected)"
 $ breaks    : list()
  ..- attr(*, "class")= chr "waiver"
 $ labels    : list()
  ..- attr(*, "class")= chr "waiver"
 $ legend    : NULL
 $ drop      : logi FALSE
 $ guide     : chr "legend"
 - attr(*, "class")= chr [1:3] "manual" "discrete" "scale"
}}
\usage{
p_value_fill
}
\description{
Palette for filling cells colored by p-value
}
\keyword{datasets}

