\docType{data}
\name{text_scale}
\alias{text_scale}
\title{Text scale- dark and light, depending on fill}
\format{\preformatted{List of 14
 $ call      : language discrete_scale(aesthetics = aesthetic, scale_name = "manual", palette = pal,      na.value = "#EEEEEE", guide = FALSE)
 $ aesthetics: chr "colour"
 $ scale_name: chr "manual"
 $ palette   :function (n)  
 $ range     :Reference class 'DiscreteRange' [package "scales"] with 1 fields
  ..$ range: NULL
  ..and 15 methods, of which 3 are possibly relevant:
  ..  initialize, reset, train
 $ limits    : NULL
 $ na.value  : chr "#EEEEEE"
 $ expand    : list()
  ..- attr(*, "class")= chr "waiver"
 $ name      : NULL
 $ breaks    : list()
  ..- attr(*, "class")= chr "waiver"
 $ labels    : list()
  ..- attr(*, "class")= chr "waiver"
 $ legend    : NULL
 $ drop      : logi TRUE
 $ guide     : logi FALSE
 - attr(*, "class")= chr [1:3] "manual" "discrete" "scale"
}}
\usage{
text_scale
}
\description{
Text scale- dark and light, depending on fill
}
\keyword{datasets}

